travis cant do test, im killing him
he can't run the server and the client at the same time, despite my best efforts. it was an interesting learning experience trying to do it and I added like 50 commits to the project trying to make it work but it wasnt meant to be. honestly, to think there'd be a way to have it run simultaneous commands. it doesnt even need to pull up two virtual machines for legit testing! even though it does do that on every commit, it pulls up another machine if the last one hasnt finished running... if you're reading this, this is the rabbit hole I fell down.